LayoutTests/fast/filesystem/workers/file-from-file-entry-sync.html

<html>
<head>
<script src="../resources/fs-worker-common.js"></script>
<script src="../resources/fs-worker-test-pre.js"></script>
</head>
<body>
<p id="description">This tests returning File from FileEntrySync in workers.</p>
<div id="console"></div>
<script>
startWorker('../resources/file-from-file-entry-sync.js');
</script>
<script src="../resources/fs-worker-test-post.js"></script>
</body>
</html>
